Bug 10743: Add styles to 'my privacy' page in OPAC
This patch does a couple things.  Most importantly it adds
a new div around the privacy rules summary so that it can
be easily changed. It also moves the start of the form to
right above the form's select box so that it's not in the
new privacy rules div.

To test:

* Apply patch
* Enable the use of privacy via the sys prefs
* Log in to the OPAC
* Visit the 'my privacy' tab
* Make sure it looks okay
* Submit the form and make sure it works
* Repeat in the other theme

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Edit: Replaced a few tabs with spaces to keep the QA script happy.
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works nicely, privacy options can still be edited.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
(cherry picked from commit 841fe4559f6264295ec4cfde6ddd6e22a151d6bf)
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>