Bug 9374: The color should manage all zero format.
The format of prices in Koha is discuted in some bugs (e.g. 9410).
The good way will be to introduce a syspref in order to deal with the
number of decimal.
The previous patch is too restrictive, we should deal with other zero
format.

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>
(cherry picked from commit c523b56bcf0eccb2feb342fe962e590d74baa749)
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>