Bug 7813: (follow-up) use descriptive text for delete link
This patch replaces the "times" symbol for the local cover image
deletion link with more descriptive text.  This was done because
the original "X" was small, is not necessarily going to be universally
understood to signify the deletion action, and because "&times;"
does not give much context to translators.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
(cherry picked from commit fe3628070a7affddb691ee589ec0e63639d0a36a)
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>